Added Eigen as a submodule, and moved submodules to externals folder.
