+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "pyBeam/CoDiPack"]
-	path = pyBeam/CoDiPack
+	path = pyBeam/externals/CoDiPack
 	url = https://github.com/SciCompKL/CoDiPack.git
+[submodule "pyBeam/externals/Eigen"]
+	path = pyBeam/externals/Eigen
+	url = https://github.com/eigenteam/eigen-git-mirror.git
